final v

+++ app/src/main/java/hse/org/ddmukhin/klavogonkiapplication/remote/data/Results.kt
@@ -1,3 +1,3 @@
 package hse.org.ddmukhin.klavogonkiapplication.remote.data
 
-data class Results(val time: Long, val speed: Long)
+data class Results(val time: Long, val speed: Double)
